+++ src/app/app.module.ts
@@ -15,6 +15,8 @@ import { SohoComponentsModule } from 'ids-enterprise-ng';
 
 import { AboutDemoComponent } from './about/about.demo';
 import { AboutNestedDemoComponent } from './about/nested-about.demo';
+import { ActionsheetDemoComponent } from './actionsheet/actionsheet.demo';
+import { ActionsheetWithTrayDemoComponent } from './actionsheet/actionsheet-with-tray.demo';
 import { AlertDemoComponent } from './alert/alert.demo';
 import { ApplicationMenuDemoComponent } from './application-menu/application-menu.demo';
 import { ApplicationMenuLazyDemoComponent } from './application-menu/application-menu-lazy.demo';
@@ -276,6 +278,8 @@ import { InputClearableDemoComponent } from './input/input-clearable.demo';
         AppComponent,
         AboutDemoComponent,
         AboutNestedDemoComponent,
+        ActionsheetDemoComponent,
+        ActionsheetWithTrayDemoComponent,
         AlertDemoComponent,
         ApplicationMenuDemoComponent,
         ApplicationMenuLazyDemoComponent,